split visualizations into their own module
