+++ tests/snap_unit_test.py
@@ -5,28 +5,7 @@ import numpy as np
 from snapanalysis.snap import snapshot
 import astropy.units as u
 import pytest
-
-CDM_TEST_SNAP_PATH = "tests/example_snaps/cdm_snaps/"
-FDM_TEST_SNAP_PATH = "tests/example_snaps/fdm_snaps/"
-
-
-@pytest.fixture
-def dm_snap():
-    s = snapshot(CDM_TEST_SNAP_PATH + "snap_000.hdf5", 1)
-    return s
-
-
-@pytest.fixture
-def fdm_snap():
-    s = snapshot(FDM_TEST_SNAP_PATH + "snap_000.hdf5", 1)
-    return s
-
-
-@pytest.fixture
-def star_snap():
-    s = snapshot(CDM_TEST_SNAP_PATH + "snap_000.hdf5", 4)
-    return s
-
+from tests.conftest import CDM_TEST_SNAP_PATH
 
 ## tests for I/O
 @pytest.mark.parametrize(
